FinnishXL/requirements.txt
# packages in environment at /home/jaina5/.conda/envs/TensorflowEnv:
#
# Name                    Version                   Build  Channel
absl-py                   0.7.1                    pypi_0    pypi
astor                     0.8.0                    pypi_0    pypi
astroid                   2.2.5                    pypi_0    pypi
attrs                     19.1.0                   pypi_0    pypi
backcall                  0.1.0                    pypi_0    pypi
blas                      1.0                         mkl  
bleach                    3.1.0                    pypi_0    pypi
ca-certificates           2019.5.15                     0  
certifi                   2019.6.16                py37_0  
cffi                      1.12.3           py37h2e261b9_0  
cudatoolkit               10.0.130                      0  
decorator                 4.4.0                    pypi_0    pypi
defusedxml                0.6.0                    pypi_0    pypi
entrypoints               0.3                      pypi_0    pypi
freetype                  2.9.1                h8a8886c_1  
gast                      0.2.2                    pypi_0    pypi
grpcio                    1.21.1                   pypi_0    pypi
h5py                      2.9.0                    pypi_0    pypi
intel-openmp              2019.4                      243  
ipykernel                 5.1.1                    pypi_0    pypi
ipython                   7.5.0                    pypi_0    pypi
ipython-genutils          0.2.0                    pypi_0    pypi
ipywidgets                7.4.2                    pypi_0    pypi
isort                     4.3.20                   pypi_0    pypi
jedi                      0.14.0                   pypi_0    pypi
jinja2                    2.10.1                   pypi_0    pypi
jpeg                      9b                   h024ee3a_2  
jsonschema                3.0.1                    pypi_0    pypi
jupyter                   1.0.0                    pypi_0    pypi
jupyter-client            5.2.4                    pypi_0    pypi
jupyter-console           6.0.0                    pypi_0    pypi
jupyter-core              4.5.0                    pypi_0    pypi
keras-applications        1.0.8                    pypi_0    pypi
keras-preprocessing       1.1.0                    pypi_0    pypi
lazy-object-proxy         1.4.1                    pypi_0    pypi
libedit                   3.1.20181209         hc058e9b_0  
libffi                    3.2.1                hd88cf55_4  
libgcc-ng                 9.1.0                hdf63c60_0  
libgfortran-ng            7.3.0                hdf63c60_0  
libpng                    1.6.37               hbc83047_0  
libstdcxx-ng              9.1.0                hdf63c60_0  
libtiff                   4.0.10               h2733197_2  
markdown                  3.1.1                    pypi_0    pypi
markupsafe                1.1.1                    pypi_0    pypi
mccabe                    0.6.1                    pypi_0    pypi
mistune                   0.8.4                    pypi_0    pypi
mkl                       2019.4                      243  
mkl_fft                   1.0.12           py37ha843d7b_0  
mkl_random                1.0.2            py37hd81dba3_0  
mock                      3.0.5                    pypi_0    pypi
nbconvert                 5.5.0                    pypi_0    pypi
nbformat                  4.4.0                    pypi_0    pypi
ncurses                   6.1                  he6710b0_1  
ninja                     1.9.0            py37hfd86e86_0  
notebook                  5.7.8                    pypi_0    pypi
numpy                     1.15.0                   pypi_0    pypi
olefile                   0.46                     py37_0  
openssl                   1.1.1c               h7b6447c_1  
pandocfilters             1.4.2                    pypi_0    pypi
parso                     0.5.0                    pypi_0    pypi
pexpect                   4.7.0                    pypi_0    pypi
pickleshare               0.7.5                    pypi_0    pypi
pillow                    6.0.0            py37h34e0f95_0  
pip                       19.1.1                   py37_0  
prometheus-client         0.7.1                    pypi_0    pypi
prompt-toolkit            2.0.9                    pypi_0    pypi
protobuf                  3.8.0                    pypi_0    pypi
ptyprocess                0.6.0                    pypi_0    pypi
pycparser                 2.19                     py37_0  
pygments                  2.4.2                    pypi_0    pypi
pylint                    2.3.1                    pypi_0    pypi
pyrsistent                0.15.2                   pypi_0    pypi
python                    3.7.3                h0371630_0  
python-dateutil           2.8.0                    pypi_0    pypi
pytorch                   1.1.0           py3.7_cuda10.0.130_cudnn7.5.1_0    pytorch
pyzmq                     18.0.1                   pypi_0    pypi
qtconsole                 4.5.1                    pypi_0    pypi
readline                  7.0                  h7b6447c_5  
send2trash                1.5.0                    pypi_0    pypi
setuptools                41.0.1                   py37_0  
six                       1.12.0                   pypi_0    pypi
sqlite                    3.28.0               h7b6447c_0  
tensorboard               1.13.1                   pypi_0    pypi
tensorflow-estimator      1.13.0                   pypi_0    pypi
tensorflow-gpu            1.13.1                   pypi_0    pypi
termcolor                 1.1.0                    pypi_0    pypi
terminado                 0.8.2                    pypi_0    pypi
testpath                  0.4.2                    pypi_0    pypi
tk                        8.6.8                hbc83047_0  
torchvision               0.3.0           py37_cu10.0.130_1    pytorch
tornado                   6.0.3                    pypi_0    pypi
traitlets                 4.3.2                    pypi_0    pypi
typed-ast                 1.4.0                    pypi_0    pypi
wcwidth                   0.1.7                    pypi_0    pypi
webencodings              0.5.1                    pypi_0    pypi
werkzeug                  0.15.4                   pypi_0    pypi
wheel                     0.33.4                   py37_0  
widgetsnbextension        3.4.2                    pypi_0    pypi
wrapt                     1.11.1                   pypi_0    pypi
xz                        5.2.4                h14c3975_4  
zlib                      1.2.11               h7b6447c_3  
zstd                      1.3.7                h0b5b093_0  
